Refactor css files into styles directory #74
