+++ pages/quiz/quiz-questions.tsx
@@ -3,9 +3,9 @@ import { NextButton } from '../../components/quiz/NextButton';
 import PreviousButton from '../../components/quiz/PreviousButton';
 import QuizButton from '../../components/quiz/QuizButton';
 import Steps from '../../components/quiz/Steps';
-import { CONTAINER_STATE } from './constants/quiz'
+import { CONTAINER_STATE } from '../../constants/quiz'
 import classes from './quiz.module.css';
-import { QuizQuestions as quizQuestions } from "./constants";
+import { QuizQuestions as quizQuestions } from "../../constants";
 import QuizOption from '../../components/quiz/QuizOption';
 import quiz from '.';
